requirements.txt
requests
beautifulsoup4
pdfplumber
pandas
google-cloud-storage
scikit-learn
xgboost
scikit-learn==1.3.0
numpy==1.26.3
streamlit==1.29.0
streamlit-option-menu==0.3.6
pandas
matplotlib
Pillow
seaborn
